refactor(chat): extract chat_manager/deps_lifecycle.rs (#292)
Task 9 of the split plan: LcDeps and its LifecycleManagerDeps impl
move into their own file, owning their own construction via a
build() helper the constructor now calls.
